Initial commit for Canvas SpeedGrader replacement project
This commit marks the start of a new project aimed at replacing Canvas SpeedGrader with a more efficient and effective grading system. The current version of the project includes a working prototype that allows for grading question-by-question, rather than student-by-student as in the original SpeedGrader. This design will make the grading process faster and more accurate, ultimately leading to a better experience for both instructors and students. As this is an initial commit, future updates will include additional features and improvements to the existing prototype.